/deleteUser endpoint changed so that it takes password as parameter if password entered matches password in db, user is deleted from database

+++ src/main/java/group13/wishlist/UserService.java
@@ -34,9 +34,9 @@ public class UserService {
     }
 
     // Delete a user by username
-    public boolean deleteUser(String username) {
+    public boolean deleteUser(String username, String password) {
         User user = userRepository.findByUsername(username);
-        if (user != null) {
+        if (user != null && passwordEncoder.matches(password, user.getPassword())) {
             userRepository.delete(user);
             return true;
         }
